+++ era_customer_success/models/__init__.py
@@ -16,3 +16,4 @@ from . import ir_attachment
 from . import cs_weekly_suggestion
 from . import cs_success_profile
 from . import cs_support_wallet
+from . import cs_value_review